Optimize source code
